pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uap-desktop"
version = "0.1.0"
description = "UAP - 复杂系统未来势态量化预测统一智能体（桌面客户端）"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pywebview>=5.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "loguru>=0.7",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "openai>=1.0.0",
    "ollama>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "setuptools>=81.0.0",
    # 本地文件 URI 依赖 milvus-lite；Windows 无官方 win 轮子，见 milvus_project_kb._milvus_lite_import_ok
    "pymilvus[milvus_lite]>=2.4.0",
]

[project.optional-dependencies]
vector = [
    "torch>=2.0",
    "sentence-transformers>=2.0",
]
legacy-db = [
    "mysql-connector-python>=8.0",
    "redis>=5.0",
    "elasticsearch>=8.0",
    "neo4j>=5.0",
    "pymongo>=4.0",
    "minio>=7.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/local/uap-desktop"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"uap.core.prompts" = ["assets/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]
